Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyodbc==4.0.39
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3